Improved considerably the application. + EventsMapperHelper: Created this new helper class. This converts a json in a EventModel  list.
+ GitHubEventsService: Now There's this service that encapsulates the logic to make the request to GitHub api and return the processed data.
It is a very simple implementation. For now only returns a list of events without putting any filter.
This service uses the http service that already existed and EventsMapperHelper.
This calls for the json GET, processes with MapperHelper and finally returns

I am listing the events list in the EditText. The next step is to put the list View.

improvements